+++ src/main/java/com/github/osisoft/dataviewsample/App.java
@@ -26,6 +26,11 @@ public class App {
     static String sampleStreamName2 = "Tank100";
     static String sampleFieldToConsolidateTo = "temperature";
     static String sampleFieldToConsolidate = "ambient_temp";
+    static String sampleFieldToAddUom1 = "pressure";
+    static String sampleFieldToAddUom2 = "temperature";
+    static String sampleFieldToSummarize = "pressure";
+    static SdsSummaryType summaryType1 = SdsSummaryType.Mean;
+    static SdsSummaryType summaryType2 = SdsSummaryType.Total;
     static Instant sampleStartTime = null;
     static Instant sampleEndTime = null;
 
@@ -198,13 +203,65 @@ public class App {
                     sampleStartTime.toString(), sampleEndTime.toString(), sampleInterval).getResponse();
             System.out.println(dataViewData);
             assert dataViewData.length() > 0 : "Error getting data view data";
+
+            // Step 12
+            System.out.println();
+            System.out.println("Step 12: Add Units of Measure Column");
+            field1 = findField(dvDataItemFieldSet.getDataFields(), FieldSource.PropertyId,
+                    sampleFieldToAddUom1);
+            field2 = findField(dvDataItemFieldSet.getDataFields(), FieldSource.PropertyId,
+                    sampleFieldToAddUom2);
+            assert field1 != null : "Error finding data field";
+            assert field2 != null : "Error finding data field";
+            System.out.println(ocsClient.mGson.toJson(field1));
+            System.out.println(ocsClient.mGson.toJson(field2));
+
+            field1.setIncludeUom(true);
+            field2.setIncludeUom(true);
+            ocsClient.DataViews.createOrUpdateDataView(namespaceId, dataView);
+
+            System.out.println("Retrieving data from the data view:");
+            dataViewData = ocsClient.DataViews.getDataViewData(namespaceId, sampleDataViewId,
+                    sampleStartTime.toString(), sampleEndTime.toString(), sampleInterval).getResponse();
+            System.out.println(dataViewData);
+            assert dataViewData.length() > 0 : "Error getting data view data";
+
+            // Step 13
+            System.out.println();
+            System.out.println("Step 13: Add Summaries Columns");
+            field1 = findField(dvDataItemFieldSet.getDataFields(), FieldSource.PropertyId,
+                    sampleFieldToSummarize);
+            assert field1 != null : "Error finding data field";
+            System.out.println(ocsClient.mGson.toJson(field1));
+
+            Field summaryField1 = new Field(field1);
+            Field summaryField2 = new Field(field1);
+
+            summaryField1.setSummaryDirection(SummaryDirection.Forward);
+            summaryField1.setSummaryType(summaryType1);
+            summaryField2.setSummaryDirection(SummaryDirection.Forward);
+            summaryField2.setSummaryType(summaryType2);
+
+            fields = new ArrayList<Field>(Arrays.asList(dvDataItemFieldSet.getDataFields()));
+            fields.add(summaryField1);
+            fields.add(summaryField2);
+            dvDataItemFieldSet.setDataFields(Arrays.copyOf(fields.toArray(), fields.size(), Field[].class));
+            ocsClient.DataViews.createOrUpdateDataView(namespaceId, dataView);
+            ocsClient.DataViews.createOrUpdateDataView(namespaceId, dataView);
+
+            System.out.println("Retrieving data from the data view:");
+            dataViewData = ocsClient.DataViews.getDataViewData(namespaceId, sampleDataViewId,
+                    sampleStartTime.toString(), sampleEndTime.toString(), sampleInterval).getResponse();
+            System.out.println(dataViewData);
+            assert dataViewData.length() > 0 : "Error getting data view data";
+
         } catch (Exception e) {
             e.printStackTrace();
             success = false;
         } finally {
-            // Step 12
+            // Step 14
             System.out.println();
-            System.out.println("Step 12: Delete sample objects from OCS");
+            System.out.println("Step 14: Delete sample objects from OCS");
             try {
                 System.out.println("Deleting data view...");
                 ocsClient.DataViews.deleteDataView(namespaceId, sampleDataViewId);
@@ -224,11 +281,11 @@ public class App {
             SdsType doubleType = new SdsType("doubleType", "", "", SdsTypeCode.Double);
             SdsType dateTimeType = new SdsType("dateTimeType", "", "", SdsTypeCode.DateTime);
 
-            SdsTypeProperty pressureDoubleProperty = new SdsTypeProperty("pressure", "", "", doubleType, false);
+            SdsTypeProperty pressureDoubleProperty = new SdsTypeProperty("pressure", "", "", doubleType, false, "bar");
             SdsTypeProperty temperatureDoubleProperty = new SdsTypeProperty(sampleFieldToConsolidateTo, "", "",
-                    doubleType, false);
+                    doubleType, false, "degree Celsius");
             SdsTypeProperty ambientTemperatureDoubleProperty = new SdsTypeProperty(sampleFieldToConsolidate, "", "",
-                    doubleType, false);
+                    doubleType, false, "degree Celsius");
             SdsTypeProperty timeDateTimeProperty = new SdsTypeProperty("time", "", "", dateTimeType, true);
 
             SdsType sdsType1 = new SdsType(sampleTypeId1, "", "", SdsTypeCode.Object,